Big refactoring & rearchitecting
This gets away from calling the main data model an "Ontology". Instead,
we deal with a *graph*, which can produce *hierarchies* based on
parent/child relations. This gets rid of the problem that kept coming up
about whether an "ontology" could only have a single root. Now, the
graph does not care about roots whatsoever, although it does track which
items have no parents, and reports those as roots for convenience.

This also gets rid of the lunr index as a built-in part of the graph.

Small thing: renames "happy path" to "preferred path".