Rewrite the Hierarchy class
This is a structural rewrite of the Hierarchy class that makes things
clearer, simpler, and allows retrieving nodes based on their path. It
was a necessary step before doing the UI work of "open a specific node
and scroll to it".

The main difference is that the primary addressing mechanism in
Hierarchy is via a new Path class. Paths point to a specific point in
the hierarchy, and have basic methods attached to them (depth(),
equals(), etc.). They can also be serialized to a string key (and can be
deserialized from that key as well).

This revamps the `buildFlatTree` method to work based on paths--
everything should be clearer now.