Get the HierarchyTree component working with selecting nodes
This is really the payoff for the last few re-architecting commits-- it
makes it possible to purposefully, imperatively open up a node in the
tree and select it.

This isn't finished, of course. Getting the data structures right and
fleshing out `openAndFocusNode` raises a lot of interaction questions.
The cool thing is that the blocks exist such that we can start asking
them and doing stuff about em! Lots of different things are possible.